diffusion/pyproject.toml
[project]
name = "car-diffusion"
version = "0.1.0"
description = "Diffusion model for car image generation using PyTorch"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.2.0",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
